ex7 enhancement
I trid not to directly eliminate errors before carrying out the nonlinear lsq.
While it is possible for receiver clock error, the ionospheric delay and the tropospheric delay. But for satellite clock error, it is not possible since each satellite will have its unique clock delay. So if we want to model that error, then for each pseudorange, we need a parameter. That will make the unknow parameters more than the measurements.